[[ REFACTOR ]] Refactor of http request module.
